+++ packages/terra-clinical-item-collection/tests/wdio/clinical-item-collection-spec.js
@@ -31,8 +31,7 @@ Terra.describeViewports('Item Collection', ['tiny', 'small', 'medium', 'large', 
       browser.click('#selected-key');
     });
 
-    // TODO: remove disabled color-contrast rule: https://github.com/cerner/terra-clinical/issues/673
-    Terra.it.validatesElement({ rules: { 'color-contrast': { enabled: false } } });
+    Terra.it.validatesElement();
 
     after(() => browser.moveToObject('#root', 0, 0));
   });