+++ packages/terra-clinical-item-display/src/ItemDisplay.jsx
@@ -49,8 +49,7 @@ const propTypes = {
 
 const defaultProps = {
   text: '',
-  // TODO: textStyle should be set to 'primary' on the next major version bump: https://github.com/cerner/terra-clinical/issues/526
-  textStyle: undefined,
+  textStyle: 'primary',
   isTruncated: false,
   isDisabled: false,
   icon: undefined,
@@ -79,7 +78,10 @@ const ItemDisplay = ({
   const textClassNames = cx([
     'text',
     { 'is-truncated': isTruncated },
-    { 'strike-through': textStyle === TextStyles.STRIKETHROUGH, [`${textStyle}`]: textStyle },
+    { [`${textStyle}`]: textStyle === TextStyles.SECONDARY },
+    { [`${textStyle}`]: textStyle === TextStyles.ATTENTION },
+    { [`${textStyle}`]: textStyle === TextStyles.STRONG },
+    { 'strike-through': textStyle === TextStyles.STRIKETHROUGH },
   ]);
 
   let displayIcon;